# fix performance issue in scenario #2966 (part 1)