Re-employing a compression state across multiple successive compression jobs
shall minimize the amount of allocation and initialization required.

This mostly matters in situations where initialization is an overwhelming task compared to compression itself.
This can happen when the amount to compress is small,
while the compression state was expecting it to be much larger,
aka, streaming mode without the benefit of a `srcSize` hindsight.
In which case, `zstd` can't rely on its automatic resource resizing capability, and initialization cost can only be made acceptable by amortizing it over a series of compressions.

This diff restores a lean-initialization optimization trick that used to be present up to `v1.4.9` and was lost in commit 980f3bbf8354edec0ad32b4430800f330185de6a .

The following measurement is taken on a core `i7-9700K` (turbo disabled) with `fullbench -b41`, using `geldings.txt` as sample (a small text file). The test corresponds to a scenario using `ZSTD_compressStream()` without the benefit of knowing the small sample size beforehand.

| level | v1.5.1 | this PR | comment
| --- | ---:| ---:| --- |
| 1 | 101 MB/s | 113 MB/s |
| 2 |  67 MB/s | 112 MB/s |
| 3 |  31 MB/s |  94 MB/s |
| 4 |  14 MB/s |  93 MB/s |

Note how important is this lean-initialization optimization for small data as the compression level increases, hence the amount of resources used increases too.

Note that this PR does not completely fix #2966,
since another heavy initialization, specific to `rowHash` mode,
is also present (and wasn't in `v1.4.9`).
This second issue will be fixed in a separate commit.

For information :
| level | v1.5.1 | this PR | comment
| --- | ---:| ---:| --- |
| 5 |   9 MB/s |  20 MB/s | `rowHash`
| ... | ... | ... | `rowHash`
|12 | 0.4 MB/s | 1.2 MB/s | `rowHash`
|13 | 0.6 MB/s |  38 MB/s |